+++ .gitignore
@@ -146,3 +146,4 @@ build/
 docs/
 
 .claude/
+CLAUDE.md
