+++ src/main/kotlin/com/yourssu/scouter/admin/AdminRequirementController.kt
@@ -7,73 +7,21 @@ import com.yourssu.scouter.recruiting.interview.application.dto.UpdateInterviewR
 import com.yourssu.scouter.recruiting.interview.business.InterviewRequirementService
 import com.yourssu.scouter.recruiting.interview.business.dto.InterviewRequirementItemDto
 import jakarta.servlet.http.HttpSession
-import java.nio.charset.StandardCharsets
-import java.security.MessageDigest
-import org.springframework.boot.context.properties.EnableConfigurationProperties
-import org.springframework.http.ResponseEntity
 import org.springframework.stereotype.Controller
 import org.springframework.ui.Model
 import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.ResponseBody
-import org.springframework.web.servlet.mvc.support.RedirectAttributes
 
 @Controller
-@RequestMapping("/admin/recruiting")
-@EnableConfigurationProperties(RecruitingAdminProperties::class)
-class RecruitingAdminController(
+@RequestMapping("/admin/recruiting/requirements")
+class AdminRequirementController(
     private val partService: PartService,
     private val interviewRequirementService: InterviewRequirementService,
-    private val properties: RecruitingAdminProperties,
 ) {
 
-    @GetMapping("")
-    fun home(session: HttpSession): String {
-        if (!isAuthorized(session)) return "admin/recruiting-login"
-        return "redirect:/admin/recruiting/parts"
-    }
-
-    @PostMapping("/auth")
-    fun auth(
-        @RequestParam password: String,
-        session: HttpSession,
-        redirectAttributes: RedirectAttributes,
-    ): String {
-        val configured = properties.password.trim()
-        if (configured.isEmpty()) {
-            redirectAttributes.addFlashAttribute("error", "비밀번호가 설정되지 않았습니다. scouter.recruiting-admin.password를 설정해주세요.")
-            return "redirect:/admin/recruiting/parts"
-        }
-        if (secureEquals(password.trim(), configured)) {
-            session.setAttribute(AdminConstants.RECRUITING_ADMIN_SESSION_KEY, true)
-            return "redirect:/admin/recruiting/parts"
-        }
-        redirectAttributes.addFlashAttribute("error", "비밀번호가 올바르지 않습니다.")
-        return "redirect:/admin/recruiting/parts"
-    }
-
-    @GetMapping("/parts")
-    fun parts(model: Model, session: HttpSession): String {
-        if (!isAuthorized(session)) return "admin/recruiting-login"
-        model.addAttribute("parts", partService.readAll().partDtos)
-        return "admin/part-assignments"
-    }
-
-    @ResponseBody
-    @PostMapping("/parts/{partId}/assignments/toggle")
-    fun toggleAssignment(
-        @PathVariable partId: Long,
-        session: HttpSession,
-    ): ResponseEntity<Unit> {
-        if (!isAuthorized(session)) return ResponseEntity.status(401).build()
-        partService.toggleAssignment(partId)
-        return ResponseEntity.ok().build()
-    }
-
-    @GetMapping("/requirements/culture-fit")
+    @GetMapping("/culture-fit")
     fun cultureFit(
         @RequestParam(required = false) semester: String?,
         model: Model,
@@ -89,7 +37,7 @@ class RecruitingAdminController(
         return "admin/requirements/culture-fit"
     }
 
-    @PostMapping("/requirements/culture-fit")
+    @PostMapping("/culture-fit")
     fun saveCultureFit(
         @RequestParam semester: String,
         @RequestParam(required = false, defaultValue = "") items: List<RequirementItemParam>,
@@ -119,7 +67,7 @@ class RecruitingAdminController(
         return "admin/requirements/culture-fit"
     }
 
-    @GetMapping("/requirements/team-fit")
+    @GetMapping("/team-fit")
     fun teamFit(
         @RequestParam(required = false) partId: Long?,
         @RequestParam(required = false) semester: String?,
@@ -138,7 +86,7 @@ class RecruitingAdminController(
         return "admin/requirements/team-fit"
     }
 
-    @PostMapping("/requirements/team-fit")
+    @PostMapping("/team-fit")
     fun saveTeamFit(
         @RequestParam partId: Long,
         @RequestParam semester: String,
@@ -171,7 +119,7 @@ class RecruitingAdminController(
         return "admin/requirements/team-fit"
     }
 
-    @GetMapping("/requirements/job-fit")
+    @GetMapping("/job-fit")
     fun jobFit(
         @RequestParam(required = false) partId: Long?,
         @RequestParam(required = false) semester: String?,
@@ -190,7 +138,7 @@ class RecruitingAdminController(
         return "admin/requirements/job-fit"
     }
 
-    @PostMapping("/requirements/job-fit")
+    @PostMapping("/job-fit")
     fun saveJobFit(
         @RequestParam partId: Long,
         @RequestParam semester: String,
@@ -225,9 +173,6 @@ class RecruitingAdminController(
 
     private fun isAuthorized(session: HttpSession): Boolean =
         session.getAttribute(AdminConstants.RECRUITING_ADMIN_SESSION_KEY) == true
-
-    private fun secureEquals(a: String, b: String): Boolean =
-        MessageDigest.isEqual(a.toByteArray(StandardCharsets.UTF_8), b.toByteArray(StandardCharsets.UTF_8))
 
     private fun List<InterviewRequirementItemDto>.toUpdateItems() =
         map { UpdateInterviewRequirementItemRequest(it.id, it.content) }
